+++ examples/event-source-mapping/main.tf
@@ -6,7 +6,6 @@ provider "aws" {
   skip_metadata_api_check     = true
   skip_region_validation      = true
   skip_credentials_validation = true
-  skip_requesting_account_id  = true
 }
 
 data "aws_availability_zones" "available" {}
@@ -27,9 +26,9 @@ module "lambda_function" {
 
   function_name = "${random_pet.this.id}-lambda-event-source-mapping"
   handler       = "index.lambda_handler"
-  runtime       = "python3.8"
+  runtime       = "python3.12"
 
-  source_path = "${path.module}/../fixtures/python3.8-app1/index.py"
+  source_path = "${path.module}/../fixtures/python-app1/index.py"
 
   event_source_mapping = {
     sqs = {
@@ -234,7 +233,7 @@ resource "aws_kinesis_stream" "this" {
 # Amazon MQ
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "~> 3.0"
+  version = "~> 5.0"
 
   name = random_pet.this.id
   cidr = local.vpc_cidr
@@ -248,7 +247,7 @@ module "vpc" {
 resource "aws_mq_broker" "this" {
   broker_name        = random_pet.this.id
   engine_type        = "RabbitMQ"
-  engine_version     = "3.10.10"
+  engine_version     = "3.12.13"
   host_instance_type = "mq.t3.micro"
   security_groups    = [module.vpc.default_security_group_id]
   subnet_ids         = slice(module.vpc.public_subnets, 0, 1)